+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 - Fixed bug in calculating output directory from command-line option.
 - Added `strip-annotations` ISP.
 - Stopped COCO/ROI/VGG image object-detection formats from writing negative annotations.
+- Fixed bug in coercions where negatives instances weren't being handled correctly.
 
 0.5.0 (2020-10-27)
 ------------------